move home page component to pages folder
